+++ src/core/server/athena/main.ts
@@ -42,6 +42,7 @@ export const DEFAULT_CONFIG = {
     BOOTUP_HOUR: 9,
     BOOTUP_MINUTE: 0,
     MINUTES_PER_MINUTE: 5,
+    USE_SERVER_TIME: true, // Overwrites the server time to use the current time of the computer it is running on.
     // Inventory Properites
     TIME_BETWEEN_INVENTORY_UPDATES: 10000,
     // Food Properties