+++ src/webview/client/hud/app.js
@@ -9,12 +9,10 @@ async function sleep(ms) {
     });
 }
 
-const audioStreams = [];
-
 const app = new Vue({
     el: '#app',
     vuetify: new Vuetify({ theme: { dark: true } }),
-    components: [actions, leaderboard, chat, help, phone, status],
+    components: [actions, leaderboard, chat, help, status],
     data() {
         return {
             show: false,
@@ -84,56 +82,12 @@ const app = new Vue({
                     audio.pause();
                 }, duration);
             }
-        },
-        audioStream(identifier, volume, startSeconds = 0) {
-            if (!('YT' in window)) {
-                return;
-            }
-
-            if (this.youtubeInfo.identifier === identifier) {
-                if (this.player.getPlayerState() === 2) {
-                    this.player.playVideo();
-                    this.player.seekTo(startSeconds, true);
-                }
-
-                this.player.setVolume(volume);
-                return;
-            }
-
-            if (!this.player) {
-                this.player = new YT.Player('player', {
-                    height: '390',
-                    width: '640',
-                    videoId: `${identifier}`,
-                    startSeconds: startSeconds,
-                    events: {
-                        onReady: (e) => {
-                            this.player.loadVideoById(identifier, startSeconds, 'large');
-                            this.player.playVideo();
-                        }
-                    }
-                });
-            } else {
-                this.player.loadVideoById(identifier, startSeconds, 'large');
-                this.player.playVideo();
-            }
-
-            this.youtubeInfo.identifier = identifier;
-        },
-        pauseStream() {
-            if (!this.player) {
-                return;
-            }
-
-            this.player.pauseVideo();
         }
     },
     mounted() {
         if ('alt' in window) {
             alt.on('leaderboard:Toggle', this.toggleLeaderboard);
             alt.on('hud:Audio3D', this.audio3D);
-            alt.on('hud:AudioStream', this.audioStream);
-            alt.on('hud:PauseStream', this.pauseStream);
             alt.on('url', this.setURL);
         } else {
             this.players = [
@@ -144,10 +98,6 @@ const app = new Vue({
             for (let i = 0; i < 5; i++) {
                 this.players = [...this.players, ...this.players];
             }
-
-            // setTimeout(() => {
-            //     this.audioStream(`KrUak31dVqc`, 25, 25);
-            // }, 1000);
 
             // setInterval(() => {
             //     this.audio3D('car_lock', 0, 1);